changes to QAConfig: - move from config.py to new qa_config.py - add to __init__.py for gamechangerml/configs - update references to this class - add docstrings
