+++ gamechangerml/configs/__init__.py
@@ -1,1 +1,2 @@
 from .s3_config import S3Config
+from .qa_config import QAConfig
